external/board/mango-rpi4-v1/overlay/opt/transport/groups/mango/support/requirements.txt
pytz
py-dateutil
pyephem
python-dateutil
pyserial
gpsd-py3
websocket-client
